Phase 12: Pure C# Codecs
Add managed image codec implementations:
- JPEG Baseline encoder/decoder with DCT
- JPEG Lossless encoder/decoder with DPCM predictors
- JPEG 2000 with DWT (5/3, 9/7), EBCOT, MQ coder
- AOT-compatible codec registration system
- BitReader/BitWriter for Huffman coding
- Color space conversion (YCbCr, RGB)

Plans 12-01 through 12-07.

+++ src/SharpDicom/Codecs/Jpeg2000/Wavelet/DwtTransform.cs
@@ -35,8 +35,7 @@ namespace SharpDicom.Codecs.Jpeg2000.Wavelet
                 throw new ArgumentException("Width and height must be positive.");
             }
 
-            // Use long multiplication to avoid overflow before comparing
-            if (data.Length < (long)width * height)
+            if (data.Length < width * height)
             {
                 throw new ArgumentException("Data buffer is too small for the specified dimensions.");
             }
@@ -87,8 +86,7 @@ namespace SharpDicom.Codecs.Jpeg2000.Wavelet
                 throw new ArgumentException("Width and height must be positive.");
             }
 
-            // Use long multiplication to avoid overflow before comparing
-            if (data.Length < (long)width * height)
+            if (data.Length < width * height)
             {
                 throw new ArgumentException("Data buffer is too small for the specified dimensions.");
             }
